+++ CHANGELOG.md
@@ -1,7 +1,6 @@
 ### Added:
-- 
+- Added range attack damage attribute
 ### Fixed:
 - 
 ### Changed:
-- Updated to mc 1.21.1
-- Removed dexterity enchantment
+-